+++ libs/estdlib/src/erlang.erl
@@ -250,7 +250,7 @@ system_flag(_Key, _Value) ->
 %%-----------------------------------------------------------------------------
 -spec md5(Data :: binary()) -> binary().
 md5(Data) when is_binary(Data) ->
-    erlang:nif_error(undefined).
+    crypto:hash(md5, Data).
 
 %%-----------------------------------------------------------------------------
 %% @param   Module Name of module