+++ tests/test.c
@@ -502,6 +502,7 @@ struct Test tests[] = {
     TEST_CASE_EXPECTED(trap_exit_flag, 1),
     TEST_CASE_COND(test_stacktrace, 0, SKIP_STACKTRACES),
     TEST_CASE(small_big_ext),
+    TEST_CASE(test_crypto),
 
     TEST_CASE(test_min_max_guard),